+++ packages/core/src/rolldown-config/plugins/clean.ts
@@ -1,5 +1,7 @@
+import { existsSync } from 'fs';
 import { relative } from 'node:path';
 
+import { findConfigFile } from '@typescript/typescript6';
 import { globSync } from 'glob';
 import type { RolldownPluginOption } from 'rolldown';
 
@@ -7,7 +9,7 @@ import { toArray } from '#utils';
 
 import { DEFAULT_DIR, WARNING_CODES } from '../constants';
 import { cleanupJS, withoutExtension } from '../helpers';
-import { findConfigFile, readTsConfig } from './typescript.config';
+import { readTsConfig } from './typescript.config';
 
 export type EndPluginOptions = {
   ignoresJS?: string | string[];
@@ -53,8 +55,12 @@ export const clean = (options: EndPluginOptions) => {
       handler: () => {
         const cwd = process.cwd();
 
-        const tsconfigPath = findConfigFile(cwd, 'tsconfig.json')!;
-        const _outDir = readTsConfig(tsconfigPath).compilerOptions.outDir;
+        const tsconfigPath = findConfigFile(
+          cwd,
+          existsSync,
+          'tsconfig.json',
+        )!;
+        const _outDir = readTsConfig(tsconfigPath).options.outDir;
         const outDir = _outDir ?? dir ?? DEFAULT_DIR;
         if (_ignoresJS.length > 0) {
           try {